added protection for members

+++ servers/website/routes/dashboard/members.coffee
@@ -1,4 +1,7 @@
 module.exports.add = (req, res, next)->
+  if req.user.id != req.publisher.owner_id
+    return next "Ownership access is required!"
+  
   email = req.body.email.toLowerCase().trim()
 
   LIBS.models.User.findOne({
@@ -37,6 +40,9 @@ module.exports.add = (req, res, next)->
 
 
 module.exports.remove_invite = (req, res, next)->
+  if req.user.id != req.publisher.owner_id
+    return res.redirect "/dashboard/#{req.publisher.key}/members"
+
   LIBS.models.UserAccess.destroy({
     where: {
       id: req.params.invite
@@ -46,5 +52,8 @@ module.exports.remove_invite = (req, res, next)->
     
 
 module.exports.remove_member = (req, res, next)->
+  if req.user.id != req.publisher.owner_id or req.params.member == req.publisher.owner_id
+    return res.redirect "/dashboard/#{req.publisher.key}/members"
+  
   req.publisher.removeMember(req.params.member).then ->
     res.redirect "/dashboard/#{req.publisher.key}/members"
